feat: introduce concept of ownership
A layer may have an "owner," which is necessarily a LayerHierarchy.

This lays the groundwork for having a set of properties (e.g. configuration
or business rules) that govern any hierarchy of layers. All layers and sub-
trees in the tree will have an owner that is responsibel for carrying and
managing those properties.